requirements.txt
pydantic[dotenv]
streamlit
requests
pydantic==1.10.7
asyncio==3.4.3
anyio==3.7.1
async-timeout==4.0.3
matplotlib==3.8.1
numpy==1.26.1
pandas==2.1.2
python-dotenv==1.0.0
seaborn==0.13.0
htbuilder==0.6.2
hydralit-components==1.0.10
plotly==5.18.0

